useState to dynamically display localStorage data in vocabulary page
